gumdrop: remove create and claim from frontend
